refactor: Mapper.kt 파일 삭제
- `Cart` 모델에 `updateQuantity` 함수 이동
- `GoodsItemExtensions.kt` 파일을 생성하여 확장 함수 이동
