+++ app/src/main/java/woowacourse/shopping/domain/model/Cart.kt
@@ -5,4 +5,6 @@ data class Cart(
     val quantity: Int,
 ) {
     val totalPrice: Int = goods.price * quantity
+
+    fun updateQuantity(newQuantity: Int): Cart = this.copy(quantity = newQuantity)
 }
